refactor: move decipher to file
